+++ test/legacy_test/test_elementwise_sub_op.py
@@ -83,31 +83,6 @@ class TestElementwiseFP16OP(TestElementwiseOp):
         self.dtype = np.float16
 
 
-class TestTensorSubSplit(OpTest):
-    def test_float16_sub(self):
-        if not core.is_compiled_with_cuda():
-            return
-
-        paddle.disable_static()
-        tensor_a = paddle.rand(shape=[5120, 4, 384, 384], dtype="float16")
-        tensor_b = paddle.rand(shape=[5120, 1, 384, 384], dtype="float16")
-        tensor_z = paddle.subtract(tensor_a, tensor_b)
-
-        in0, in1 = paddle.split(tensor_a, num_or_sections=2, axis=1)
-        (
-            out0,
-            out1,
-        ) = paddle.split(tensor_z, num_or_sections=2, axis=1)
-
-        split_add0 = paddle.subtract(tensor_b, in0)
-        split_add1 = paddle.subtract(tensor_b, in1)
-
-        result1 = paddle.any(paddle.equal(out0, split_add0), [0, 1, 2, 3])
-        result2 = paddle.any(paddle.equal(out1, split_add1), [0, 1, 2, 3])
-        np.testing.assert_equal(result1.numpy(), True)
-        np.testing.assert_equal(result2.numpy(), True)
-
-
 @unittest.skipIf(
     not core.is_compiled_with_cuda()
     or not core.is_bfloat16_supported(core.CUDAPlace(0)),
